Initial commit
Notes: There are several very important TODOs on things I dont care to spend more time on for a coding assessment, for example including more tests (I included some for an example, but did not include all), and figuring out @retryable. Instead of retrable, I implemented a very quick, and poor version that seems sufficient, but would need to be replaced in a real service.